Fix parallel data providers
Since `org.testng.ITestNGMethod#getId` is not reliable, the execution
listener now uses the ITestResult as a key to find the correct
invocation and does not eagerly report the method container as finished
but relies on onAfterClass to do so. This approach is also used by
Gradle's listener implementation so it should be robust.

Fixes #4.